% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_qmm.R
\name{normalise_qmm}
\alias{normalise_qmm}
\title{Caculate the Normalised Leptokurtic Diffusion from a q-Space Myelin Map}
\usage{
normalise_qmm(qmm, outfile = NULL, write = TRUE, verbose = TRUE)
}
\arguments{
\item{qmm}{Nifti object of a processed myelinmap}

\item{outfile}{Output file name}

\item{write}{If = TRUE, write a Nifti file for the NLD in the current directory}

\item{verbose}{If = TRUE show log messages}
}
\value{
A calculated NLD nifti object
}
\description{
Caculate the Normalised Leptokurtic Diffusion from a q-Space Myelin Map
}
\author{
R.C.S
}
